[pdex] move to new file for rpc interface
